Remove root restriction. It breaks Cromwell GCP in a weird way
cnvkitBatch will reindex a .bam file if its .bai was created before
the .bam. Locally this isn't a problem but in GCP when the files are
localized, the .bai downloads first. According to cnvkit the solution
is to `touch` the .bai file. In the GCP VM created by Cromwell for
this task, however, permissions to do this are denied because
Cromwell's /cromwell_root/ directory requires root permissions to
write. cnvkit Docker image strips those away, preventing the `touch`
from working.

In version 0.9.5 the GCP instance happily creates the index, but in
0.9.7+ the script throws an error. We hoped to avoid this error by
avoiding the reindex, but were prevented by the permissions
issue. This should resolve that.